retrieval/pyproject.toml
[project]
name = "prod-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "datasets>=4.6.1",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.17",
    "langchain-google-genai>=4.2.1",
    "langchain-huggingface>=1.2.1",
    "langchain-postgres>=0.0.17",
    "langchain-text-splitters>=1.1.1",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pymupdf>=1.27.1",
    "pymupdf4llm>=0.3.4",
    "pypdf>=6.7.5",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "ragas>=0.4.3",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.3",
]
